+++ app/controllers/sessions_controller.rb
@@ -1,7 +1,7 @@
 class SessionsController < ApplicationController
   def create
     auth = request.env["omniauth.auth"]
-    account = ::Account.from_omniauth(auth)
+    account = BookingSyncEngine.account_model.from_omniauth(auth, request.host)
     account_authorized(account)
     redirect_to after_bookingsync_sign_in_path
   end