+++ lib/bookingsync/engine.rb
@@ -1,22 +1,30 @@
-require 'omniauth'
-require 'omniauth-bookingsync'
-require 'bookingsync-api'
+require "omniauth"
+require "omniauth-bookingsync"
+require "bookingsync-api"
 
 module BookingSync
   class Engine < ::Rails::Engine
     initializer "bookingsync.add_omniauth" do |app|
       app.middleware.use OmniAuth::Builder do
         provider :bookingsync,
-          ENV['BOOKINGSYNC_APP_ID'],
-          ENV['BOOKINGSYNC_APP_SECRET'],
-          scope: ENV['BOOKINGSYNC_SCOPE'],
+          BookingSyncEngine.support_multi_applications? ? nil : ENV["BOOKINGSYNC_APP_ID"],
+          BookingSyncEngine.support_multi_applications? ? nil : ENV["BOOKINGSYNC_APP_SECRET"],
+          scope: ENV["BOOKINGSYNC_SCOPE"],
           setup: -> (env) {
-            if url = ENV['BOOKINGSYNC_URL']
-              env['omniauth.strategy'].options[:client_options].site = url
+            if url = ENV["BOOKINGSYNC_URL"]
+              env["omniauth.strategy"].options[:client_options].site = url
             end
-            env['omniauth.strategy'].options[:client_options].ssl = {
-              verify: ENV['BOOKINGSYNC_VERIFY_SSL'] != 'false'
+            env["omniauth.strategy"].options[:client_options].ssl = {
+              verify: ENV["BOOKINGSYNC_VERIFY_SSL"] != "false"
             }
+
+            if BookingSyncEngine.support_multi_applications?
+              credentials = BookingSync::Engine::CredentialsResolver.new(env["HTTP_HOST"]).call
+              if credentials.valid?
+                env["omniauth.strategy"].options[:client_id] = credentials.client_id
+                env["omniauth.strategy"].options[:client_secret] = credentials.client_secret
+              end
+            end
           }
       end
     end
@@ -67,21 +75,22 @@ module BookingSync
     # The ENV variables used for configuration are described in {file:README.md}.
     #
     # @return [OAuth2::Client] configured OAuth client
-    def self.oauth_client
+    def self.oauth_client(client_id: ENV["BOOKINGSYNC_APP_ID"], client_secret: ENV["BOOKINGSYNC_APP_SECRET"])
       client_options = {
-        site: ENV['BOOKINGSYNC_URL'] || 'https://www.bookingsync.com',
+        site: ENV["BOOKINGSYNC_URL"] || 'https://www.bookingsync.com',
         connection_opts: { headers: { accept: "application/vnd.api+json" } }
       }
       client_options[:ssl] = { verify: ENV['BOOKINGSYNC_VERIFY_SSL'] != 'false' }
-      OAuth2::Client.new(ENV['BOOKINGSYNC_APP_ID'], ENV['BOOKINGSYNC_APP_SECRET'],
-        client_options)
+      OAuth2::Client.new(client_id, client_secret, client_options)
     end
 
-    def self.application_token
-      oauth_client.client_credentials.get_token
+    def self.application_token(client_id: ENV["BOOKINGSYNC_APP_ID"], client_secret: ENV["BOOKINGSYNC_APP_SECRET"])
+      oauth_client(client_id: client_id, client_secret: client_secret).client_credentials.get_token
     end
   end
 end
 
+require "bookingsync/engine/application_credentials"
+require "bookingsync/engine/credentials_resolver"
 require "bookingsync/engine/api_client"
-require "bookingsync/engine/model"
+require "bookingsync/engine/models"