+++ lib/bookingsync/engine/models.rb
@@ -1,0 +1,11 @@
+module BookingSync
+  class Engine
+    module Models
+    end
+  end
+end
+
+require "bookingsync/engine/models/base_account"
+require "bookingsync/engine/models/account"
+require "bookingsync/engine/models/multi_applications_account"
+require "bookingsync/engine/models/application"