+++ lib/bookingsync/engine/models/application.rb
@@ -1,0 +1,9 @@
+module BookingSync::Engine::Models::Application
+  extend ActiveSupport::Concern
+
+  included do
+    validates :host, presence: true, uniqueness: true
+    validates :client_id, presence: true, uniqueness: true
+    validates :client_secret, presence: true, uniqueness: true
+  end
+end