+++ spec/dummy/app/models/application.rb
@@ -1,0 +1,3 @@
+class Application < ActiveRecord::Base
+  include BookingSync::Engine::Models::Application
+end