+++ spec/dummy/db/migrate/20140522110326_create_accounts.rb
@@ -1,4 +1,4 @@
-class CreateAccounts < ActiveRecord::Migration
+class CreateAccounts < ActiveRecord::Migration[5.2]
   def change
     create_table :accounts do |t|