+++ spec/dummy/db/migrate/20181130063104_add_credentials_fields_to_applications.rb
@@ -1,0 +1,10 @@
+class AddCredentialsFieldsToApplications < ActiveRecord::Migration[5.2]
+  def change
+    add_column :applications, :host, :string, null: false
+    add_index :applications, :host, unique: true
+    add_column :applications, :client_id, :text, null: false
+    add_index :applications, :client_id, unique: true
+    add_column :applications, :client_secret, :text, null: false
+    add_index :applications, :client_secret, unique: true
+  end
+end