+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Master
 
+* Add support for multi application setup
+* BREAKING CHANGE: Account models method `from_omniauth` now takes the host as second argument
+* BREAKING CHANGE: Rename BookingSync::Engine::Model in BookingSync::Engine::Models::Account
+* BREAKING CHANGE: Drop support of Ruby 2.2
 * Relax omniauth-bookingsync requirements to >= 0.5.0
 
 # 3.0.2 (2017-11-30)